implemented caching for videos
