+++ renderer/src/components/EditorPage/Reference/TranslationHelpsImageCard.js
@@ -2,17 +2,30 @@
 // import PropTypes from 'prop-types';
 import React, { useEffect, useState } from 'react';
 import i18n from '../../../translations/i18n';
-const { net } = require('electron');
-// import {
-//   useContent,
-// } from 'translation-helps-rcl';
-// import localForage from 'localforage';
-// import LoadingScreen from '@/components/Loading/LoadingScreen';
-// import ReferenceCard from './ReferenceCard';
-// import * as logger from '../../../logger';
-// import packageInfo from '../../../../../package.json';
-// import TabSelector from './TabSelector';
-// import './TranslationHelpsImageCard.css'; // Include CSS styles
+
+const VideoPlayer = ({ videoUrl }) => {
+  const [videoSrc, setVideoSrc] = useState(videoUrl); // Default to remote video URL
+
+  useEffect(() => {
+    const loadVideo = async () => {
+      // Check if video is already cached
+      const cachedVideoPath = await global.ipcRenderer.invoke('is-video-cached', videoUrl);
+      if (cachedVideoPath) {
+        setVideoSrc(`file://${cachedVideoPath}`); // Use cached video path
+      } else {
+        // If not cached, download and cache the video
+        const downloadedPath = await global.ipcRenderer.invoke('download-and-cache-video', videoUrl);
+        setVideoSrc(`file://${downloadedPath}`); // Set downloaded video path
+      }
+    };
+
+    loadVideo();
+  }, [videoUrl]);
+
+  return (
+    <video src={videoSrc} controls className="gallery-video" />
+  );
+};
 
 export default function TranslationHelpsImageCard({
   verse,
@@ -25,56 +38,7 @@ export default function TranslationHelpsImageCard({
 
   const convertToFileUrl = (path) => `file://${path}`;
 
-  const CACHE_DIR = path.join(folderPath);
-
-  const isVideoCached = (videoUrl) => {
-    const videoFileName = path.basename(videoUrl);
-    const localVideoPath = path.join(CACHE_DIR, videoFileName);
-    return fs.existsSync(localVideoPath) ? localVideoPath : null;
-  };
-
-  const downloadAndCacheVideo = (videoUrl) => {
-    return new Promise((resolve, reject) => {
-      const videoFileName = path.basename(videoUrl);
-      const localVideoPath = path.join(CACHE_DIR, videoFileName);
-  
-      const request = net.request(videoUrl);
-      const fileStream = fs.createWriteStream(localVideoPath);
-  
-      request.on('response', (response) => {
-        response.pipe(fileStream);
-        fileStream.on('finish', () => {
-          resolve(localVideoPath);
-        });
-      });
-  
-      request.on('error', (err) => {
-        reject(err);
-      });
-  
-      request.end();
-    });
-  };
-
-  const loadVideo = async (videoUrl) => {
-    // Check if video is already cached
-    let localVideoPath = isVideoCached(videoUrl);
-    
-    if (!localVideoPath) {
-      // If not cached, download and cache the video
-      console.log(`Downloading and caching video: ${videoUrl}`);
-      localVideoPath = await downloadAndCacheVideo(videoUrl);
-    }
-  
-    return localVideoPath;
-  };
-  
-  
-
-
-  if (!fs.existsSync(CACHE_DIR)) {
-    fs.mkdirSync(CACHE_DIR);
-  }
+  // const CACHE_DIR = path.join(folderPath, 'cached_videos');
 
   /**
    * Function to get image paths from a TSV file based on book code, chapter, and verse.
@@ -93,7 +57,6 @@ export default function TranslationHelpsImageCard({
     if (!fs.existsSync(filePath)) {
       return [];
     }
-    
     
     // Read the TSV file
     const fileContent = fs.readFileSync(filePath, 'utf-8');
@@ -167,12 +130,13 @@ export default function TranslationHelpsImageCard({
               <div className="image-name">{imagePath[0]}</div>
               {/* Check if the path is a video (e.g., ends with .mp4) */}
               {imagePath[1].match(/\.(mp4|webm|ogg)$/i) ? (
-                <video
-                  key={index}
-                  src={imagePath[1]}
-                  controls
-                  className="gallery-video"
-                />
+                // <video
+                //   key={index}
+                //   src={imagePath[1]}
+                //   controls
+                //   className="gallery-video"
+                // />
+                <VideoPlayer key={index} videoUrl={imagePath[1]} />
               ) : (
                 <img
                   key={index}
